Adding support for Web Scenario creation.
Motivation:
The goal of this is to be able to configure a Zabbix server. Our current
needs are to have a docker image that can be run as a post deploy hook
to configure web scenarios (http checks) of a number of http services
that expose a ping url.

Modifications:
Added fh-zabbix.js module and run.sh which uses the module.

Result:
An buildable Docker image which can be run as a post deploy hook when
deploying a Zabbix container.

JIRA:
https://issues.jboss.org/browse/RHMAP-2765